langchain-pt2/requirements.txt
openai>=1.13.3
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.1
langchainhub>=0.1.15
pandas>=2.3.0
numpy>=2.1.0
pydantic>=2.0.0